+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+### Changed
+- Parse labels to String on initialization
 
 ## 1.0.0 - 2021-06-24
 ### Added